+++ src/angular/main.js
@@ -1,47 +1,10 @@
-System.register(["@stratus/angular/polyfills", "@stratus/core/dom", "@angular/core", "@angular/platform-browser-dynamic", "@stratus/angular/app.module", "@stratus/core/environment"], function (exports_1, context_1) {
+System.register(["@stratus/angular/polyfills", "@angular/core", "@angular/platform-browser-dynamic", "@stratus/angular/app.module", "@stratus/core/environment"], function (exports_1, context_1) {
     "use strict";
-    var dom_1, core_1, platform_browser_dynamic_1, app_module_1, environment_1, boot;
+    var core_1, platform_browser_dynamic_1, app_module_1, environment_1;
     var __moduleName = context_1 && context_1.id;
-    function angularBoot() {
-        if (boot) {
-            console.log('stopped angular boot attempt after successful boot.');
-            return;
-        }
-        const sa = [
-            'sa-selector',
-            'sa-tree',
-            'quill-editor'
-        ];
-        let detected = false;
-        sa.forEach(component => {
-            if (detected) {
-                return;
-            }
-            const elements = document.getElementsByTagName(component);
-            if (!elements || !elements.length) {
-                return;
-            }
-            detected = true;
-        });
-        if (!detected) {
-            setTimeout(() => {
-                angularBoot();
-            }, 3000);
-            return;
-        }
-        boot = true;
-        platform_browser_dynamic_1.platformBrowserDynamic().bootstrapModule(app_module_1.AppModule)
-            .then(module => {
-            console.log('@stratus/angular initialized successfully!');
-        })
-            .catch(err => console.error('@stratus/angular failed to initialize!', err));
-    }
     return {
         setters: [
             function (_1) {
-            },
-            function (dom_1_1) {
-                dom_1 = dom_1_1;
             },
             function (core_1_1) {
                 core_1 = core_1_1;
@@ -60,8 +23,11 @@ System.register(["@stratus/angular/polyfills", "@stratus/core/dom", "@angular/co
             if (null === environment_1.cookie('env')) {
                 core_1.enableProdMode();
             }
-            boot = false;
-            dom_1.DOMComplete().then(() => angularBoot());
+            platform_browser_dynamic_1.platformBrowserDynamic().bootstrapModule(app_module_1.AppModule)
+                .then(module => {
+                console.log('@stratus/angular initialized successfully!');
+            })
+                .catch(err => console.error('@stratus/angular failed to initialize!', err));
         }
     };
 });
